Crate xadrez

Xadrez

Xadrez is a Rust library for working with chess games. It provides functionality for parsing FEN strings, generating legal moves, searching for the best possible moves, and evaluating the position of the board.

Features

Parsing FEN strings into chess positions

Generating legal moves for a given chess position

Searching for the best possible move using different algorithms

Evaluating the position of the board using different metrics

Supports all standard chess rules, including en passant, castling, and pawn promotion

Written entirely in Rust, providing a safe and efficient implementation

Well-documented API and easy-to-use interface

Getting Started

To use Xadrez in your Rust project, simply add the following to your Cargo.toml file:

[dependencies]
xadrez = { git = "https://github.com/jovialen/xadrez" }

Then, import the library in your code using:

use xadrez::prelude::*;

Example Usage

Here’s an example of using Xadrez to generate all legal moves for a given position:

use xadrez::prelude::*;

let chessboard = Chessboard::default();
let moves = position.moves();

for mv in moves {
    println!("{}", mv);
}

This will output the following legal moves:

a2a3
a2a4
b2b3
b2b4
c2c3
c2c4
d2d3
d2d4
e2e3
e2e4
f2f3
f2f4
g2g3
g2g4
h2h3
h2h4
b1a3
b1c3
f1g3
f1h3

License

Xadrez is licensed under the Apache-v2.0 License.

Modules

board

Chessboard structures.

builder

Chessboard Builder.

error

Error types of the chess engine.

evaluation

Evaluation of the board.

fen

FEN structures.

move

Chess move structures.

piece

Chess piece structures.

prelude

The Xadrez prelude module provides a convenient way to bring common Xadrez types and traits into scope. It re-exports the most commonly used items from the library modules.

search

Move searching.

square

Chessboard squares.