Struct xadrez::board::Chessboard

pub struct Chessboard { /* private fields */ }

A game of Chess.

Implementations

impl Chessboard

pub fn new(fen: &str) -> Self

Constructs a new Chessboard from a Forsyth-Edwards Notation (FEN) string.

If you are uncertain whether the given FEN string is valid, use the from_fen function instead, which returns a Result.

Arguments

fen - A FEN string representing the position of the chessboard.

Examples

use xadrez::prelude::*;

let starting_fen = "rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1";

let board = Chessboard::from_fen(starting_fen).unwrap();

assert_eq!(board.side_to_move(), Side::White);

Panics

This function panics if the given FEN string is invalid.

pub fn from_fen(fen: &str) -> Result<Self, ParseFenError>

Creates a new Chessboard from a Forsyth-Edwards Notation (FEN) string.

The FEN string describes the starting position of the Chessboard and includes information about which player is to move next, castling rights, en passant squares, and the halfmove clock and fullmove clock for move tracking.

Examples

use xadrez::prelude::*;

let starting_fen = "rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1";

let board = Chessboard::from_fen(starting_fen).unwrap();

assert_eq!(board.side_to_move(), Side::White);

Errors

If the given FEN string is invalid, a ParseFenError will be returned.

pub fn moves(&self) -> Vec<Move>

Returns a vector of all possible legal moves in the current position.

pub fn is_legal(&self, m: Move) -> bool

Checks if the given move is legal on the chessboard.

Examples

use xadrez::prelude::*;

let board = Chessboard::new("rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1");
let e2e4 = Move::new(Square::E2, Square::E4).unwrap();
assert!(board.is_legal(e2e4));

let invalid_move = Move::new(Square::E3, Square::E4).unwrap();
assert!(!board.is_legal(invalid_move));

pub fn pieces(&self) -> Vec<(Square, Piece)>

Returns a vector of tuples containing all the pieces on the board along with their squares.

pub fn squares(&self) -> [(Square, Option<Piece>); 64]

Get an array of the squares on the board and the piece on it.

Returns an array representing the chessboard, with each element containing a tuple of the square and an optional piece occupying it. If a square is empty, its corresponding tuple contains a None value for the piece.

Examples

use xadrez::prelude::*;

let board = Chessboard::new("rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1");
let squares = board.squares();

assert_eq!(squares[0], (Square::A1, Piece::new(Side::White, PieceKind::Rook)));
assert_eq!(squares[1], (Square::B1, Piece::new(Side::White, PieceKind::Knight)));
assert_eq!(squares[2], (Square::C1, Piece::new(Side::White, PieceKind::Bishop)));
// and so on...

pub fn game_state(&self) -> GameState

Returns the current state of the game, which can be one of three possible values:

GameState::Checkmate - The game has ended due to one player being in checkmate.

GameState::Draw(reason) - The game has ended in a draw, with a specific DrawReason given.

GameState::Playing - The game is still ongoing.

Examples

let board = Chessboard::new("rnbqkbnr/pppppppp/8/8/8/8/PPPPPPPP/RNBQKBNR w KQkq - 0 1");
assert_eq!(board.game_state(), GameState::Playing);

A draw can occur for different reasons, as represented by the DrawReason enum:

DrawReason::Stalemate: The player whose turn it is to move has no legal moves, but is not in check.

DrawReason::Rule50: The game is drawn because 50 moves by each player have been played without a pawn move or capture.

pub fn in_check(&self) -> bool

Returns true if the current side to move is in check, and false otherwise.

pub fn side_to_move(&self) -> Side

Returns the Side enum value representing the current side to move, either Side::White or Side::Black.

pub fn evaluate(&self) -> Score

Evaluates the current position of the chessboard and returns a Score struct.

The evaluation function uses either HCE (Handcrafted Evaluation Function) or NNUE (Efficiently Updatable Neural Network Evaluation Function) depending on whether the nnue feature is enabled at compile time. NNUE is a neural network-based evaluation function that has been shown to be much more efficient than HCE, and is used if the feature is enabled. In testing, this NNUE implementation has a better early and midgame, but struggles more with the endgame.

pub fn set_position(&mut self, fen: &str) -> Result<(), ParseFenError>

Sets the position on the chessboard to the one specified by the Forsyth-Edwards Notation (FEN) string fen.

If the FEN string is invalid, the position remains unchanged and an error is returned.

Arguments

fen: A string slice containing the FEN notation of the desired position.

Errors

Returns a ParseFenError if the given FEN string is invalid.

Examples

use xadrez::prelude::*;

let mut board = Chessboard::default();
assert!(board.set_position("r1bqkbnr/pppppppp/2n5/1B6/4P3/8/PPPP1PPP/RNBQK1NR w KQkq - 0 2").is_ok());

Remarks

The previously stored position can be restored by calling the undo function.

pub fn make_move(&mut self, m: Move) -> Result<Move, MoveError>

Makes a move on the chessboard.

If the move isn’t legal, no changes will occur on the board and the function will return an error. If the move is legal, the function will return the move that was made.

Arguments

m - The Move to be made.

Errors

Returns a MoveError if the move isn’t legal.

pub fn undo(&mut self)

Undo the most recent move made on the board.

If there is no move to undo, the function returns without modifying the current position.

If a position was changed using the set_position function, the last position will also be restored.

pub fn search(
&self,
limits: SearchLimits
) -> Result<(Move, SearchData), SearchError>

Searches for the best move in the current position within the specified search limits.

Arguments

limits - The search limits to be used during the search.

Examples

let board = Chessboard::default();
let limits = SearchLimits::from_duration(Duration::from_secs(10));
let result = board.search(limits);
match result {
    Ok((m, data)) => {
        println!("best move found: {} ({})", m, data.score);
    }
    Err(e) => {
        println!("search error: {}", e);
    }
}

Errors

Will return an SearchError if the search cannot be completed.

pub fn perft(&self, depth: usize, print: bool) -> usize

Returns the number of leaf nodes in the game tree for the current position at the given depth. This function is primarily used for testing and debugging purposes.

The print flag indicates whether or not to print the result of each move calculation to the console. The result is the total number of leaf nodes found in the game tree at the given depth.

Arguments

depth - The depth at which to calculate the leaf nodes in the game tree.

print - Whether or not to print the result of each move calculation to the console.

Examples

use xadrez::Chessboard;

let board = Chessboard::default();
let nodes = board.perft(4, false);
assert_eq!(nodes, 197281);