Module xadrez::prelude

The Xadrez prelude module provides a convenient way to bring common Xadrez types and traits into scope. It re-exports the most commonly used items from the library modules.

Usage:

use xadrez::prelude::*;

This will bring the most commonly used types and traits into scope. Note that not all types and traits are exported by this module, so some items may need to be brought into scope manually.

For more information on the available types and traits, see the corresponding modules.

Re-exports

pub use crate::board::Chessboard;

pub use crate::board::DrawReason;

pub use crate::board::GameState;

pub use crate::builder::BoardBuilder;

pub use crate::piece::Piece;

pub use crate::piece::PieceKind;

pub use crate::piece::Side;

pub use crate::move::Move;

pub use crate::search::SearchData;

pub use crate::search::SearchLimits;

pub use crate::square::Square;